onboarding-api/requirements.txt
Flask
gunicorn
boto3

# OpenTelemetry (app-side)
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-botocore
opentelemetry-api
opentelemetry-sdk